+++ src/scenes/intro.js
@@ -1,5 +1,7 @@
 import { k } from "../game.js"
 import "./level-01.js"
+import "./level-02.js"
+import "./level-03.js"
 
 /**
  * Die Funktion `scene` kommt von Kaboom, und erstellt uns einen