+++ src/scenes/level-02.js
@@ -2,7 +2,8 @@ import { k, addGeneralGameLogic } from "../game.js"
 import { generateMapRPG } from "../map.js"
 import { loadKeyboardRPG } from "../keyboard.js"
 
-import "./finish.js"
+import "./level-03.js"
+import "./lose.js"
 
 /**
  * Szene für das Level 2.
@@ -19,12 +20,18 @@ k.scene("level-02", async () => {
 
   k.onCollide("player", "cave", (player) => {
     if (player.hasFlower === true) {
-      k.go("finish")
     }
+    k.go("level-03")
   })
 
   k.onCollide("player", "flower", (player, flower) => {
     flower.destroy()
     player.hasFlower = true
   })
+})
+k.onUpdate(() => {
+  const player = k.get("player")[0]
+  if (player.pos.y > 720) {
+    k.go("lose")
+  }
 })