+++ src/scenes/lose.js
@@ -1,6 +1,7 @@
 import { k } from "../game.js"
 
 import "./level-01.js"
+import "./level-02.js"
 
 /**
  * Diese Szene  wird verwendet wenn das Spiel verloren ist, also wenn der